Add a Visual Studio project file.
This unifies the source code structure of EnlyzeDbfLib with our other library projects.

This project file enforces a source code structure for the application: EnlyzeDbfLib needs to be added to the top-level "src" directory, with EnlyzeWinCompatLib next to it.
On the plus side, we can actually have a project file for EnlyzeDbfLib and don't need to add its source files manually.